Remove preciseTime, infer date precision from dateString
